+++ nni/nas/mutable/__init__.py
@@ -1,4 +1,4 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-from .engine import *
+from .mutator import *
