+++ nni/nas/utils/__init__.py
@@ -1,4 +1,5 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-from nni.nas.execution.pytorch.converter.graph_gen import *
+from .misc import *
+from .serializer import *
