+++ nni/retiarii/codegen/pytorch.py
@@ -1,4 +1,4 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-from nni.nas.execution.pytorch.graph.codegen import *
+from nni.nas.execution.pytorch.codegen import *
