+++ nni/retiarii/codegen/tensorflow.py
@@ -1,4 +1,4 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-from nni.nas.execution.tensorflow.graph.codegen import *
+from nni.nas.execution.tensorflow.codegen import *
