+++ nni/retiarii/converter/op_types.py
@@ -1,4 +1,4 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-from nni.nas.execution.pytorch.graph.converter.op_types import *
+from nni.nas.execution.pytorch.converter.op_types import *
