+++ nni/retiarii/converter/utils.py
@@ -1,4 +1,4 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-from nni.nas.execution.pytorch.graph.converter.utils import *
+from nni.nas.execution.pytorch.converter.utils import *
